refactor utils dotacao e validacao dotacao
